Make code Zig 0.91 compatible
It seems like the code for allocators and general memory handling has changed.

We no longer pass around pointers to allocators but actually duplicate the allocator structs.

Duplication and freeing of dictionary entries are different. We get a pointer to dictionary entries. I suppose that is so that you can modify entries when you iterate over them. OTOH that is a bit weird. Modifying keys during iteration seems risky.